move service.py into plugin_service folder
