+++ .env.sample
@@ -1,3 +1,5 @@
 NTS_HOST=nts.host.example.com
 NTS_PORT=8000
+NTS_KEY=
 JUPYTER_TOKEN=please_add_a_strong_password_here
+REDIS_CHANNEL=random_string_channel_name
