+++ .gitignore
@@ -16,3 +16,5 @@ cookbook/.jupyter
 cookbook/.local
 cookbook/.ipython
 cookbook/.nanome_lib
+.env*
+__pycache__
