+++ docker-compose.yml
@@ -15,17 +15,14 @@ services:
       - REDIS_PASSWORD=${REDIS_PASSWORD:-}
       - NTS_HOST=${NTS_HOST:-}
       - NTS_PORT=${NTS_PORT:-}
+      - NTS_KEY=${NTS_KEY:-}
       - ARGS=${ARGS:-}
       - JUPYTER_TOKEN=${JUPYTER_TOKEN:-change_me}
-      # - PERMANENT_CHANNEL=a5100bcd-3b33-48dc-a8b0-67f88182f24e
+      - PERMANENT_CHANNEL=a5100bcd-3b33-48dc-a8b0-67f88182f24e
     depends_on:
       - redis
     volumes:
       - ./plugin_service:/app
-      # Note: service.py and interface.py inserted into nanome-lib because for 
-      # serialization purposes, they need to be accessible by both the plugin and the client
-      - ./service.py:/opt/conda/lib/python3.7/site-packages/nanome/service.py
-      - ./interface.py:/opt/conda/lib/python3.7/site-packages/nanome/interface.py
     command: python run.py -a ${NTS_HOST} -p ${NTS_PORT} -r -v
   
   cookbook:
@@ -41,7 +38,6 @@ services:
       - 8888:8888
     volumes:
       - ./cookbook:/app
-      - ./service.py:/opt/conda/lib/python3.9/site-packages/nanome/service.py
       - ./interface.py:/opt/conda/lib/python3.9/site-packages/nanome/interface.py
     tty: true
     stdin_open: true
